requirements.txt
jupyterlab==3.4.5
jupyterlab_vim==0.15.1
ipyleaflet==0.17.1
folium==0.12.1.post1
gpxpy==1.5.0
ipywidgets==7.7.2
pillow==9.2.0
numpy==1.23.2
opencv-python==4.6.0.66
voila==0.3.6
exif==1.3.5
xdg==5.1.1
sqlalchemy==1.4.42
